src/requirements.txt
# Azure Functions and Core
azure-functions==1.17.0
# azure-functions-worker - NOT NEEDED for dev/CI (Azure runtime provides it)
# Latest PyPI version is 1.1.9 which requires old grpcio~=1.33.1 (incompatible with Python 3.11)

# Azure SDK
azure-identity==1.15.0
azure-data-tables==12.4.4
azure-storage-blob==12.19.0

# gRPC - Explicitly pin to versions with Python 3.11 wheels
# (Transitive dependency from azure-identity)
grpcio>=1.60.0,<2.0.0
grpcio-tools>=1.60.0,<2.0.0

# Not needed for MVP (uncomment when implementing)
# azure-keyvault-secrets==4.7.0
azure-storage-queue==12.8.0  # Required by setup-local.sh for queue creation
# azure-monitor-query==1.2.0  # Pulls in grpcio - use when implementing Application Insights

# Microsoft Graph
msal==1.25.0
requests==2.31.0

# Data Validation and Processing
pydantic==2.10.3
email-validator==2.1.0  # Required by Pydantic EmailStr
python-ulid==2.2.0

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1

# Type Checking and Linting
mypy==1.7.0
types-requests==2.31.0

# Security Scanning
bandit==1.7.5